requirements.txt
asgiref==3.3.4
Django==3.2.4
django-markdown2==0.3.1
django-markdownx==3.0.1
django-pygments==0.4
django-quill-editor==0.1.22
Markdown==3.3.4
markdown2==2.4.0
Pillow==8.2.0
Pygments==2.9.0
pytz==2021.1
sqlparse==0.4.1
